Update plugin validations options
