+++ .eslintrc
@@ -1,18 +1,25 @@
 {
-  "ecmaVersion": 6,
-  "ecmaFeatures": {
-    "jsx": false,
-    "modules": false
+  "globals": {
+    "jQuery": true,
+    "$": true
   },
   "env": {
     "browser": true,
     "node": true,
-    "es6": false,
+    "es6": true,
     "qunit": true
   },
   "rules": {
     "quotes": [2, "single"],
     "strict": "off"
   },
+  "parserOptions": {
+    "sourceType": "module",
+    "ecmaVersion": 6,
+    "ecmaFeatures": {
+      "jsx": false,
+      "modules": false
+    }
+},
   "plugins": []
 }
