+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "videojs-concurrence-limiter",
-  "version": "0.5.1",
+  "version": "0.6.0",
   "description": "A Video.js plugin for concurrency control on played contents",
   "repository": {
     "type": "git",
@@ -12,39 +12,39 @@
   "homepage": "https://github.com/toolbox-tve/videojs-concurrence-limiter",
   "main": "es5/plugin.js",
   "dependencies": {
-    "jquery": "^2.2.3",
-    "video.js": "^5.6.0"
+    "jquery": "^2.2.4",
+    "video.js": "^5.20.4"
   },
   "devDependencies": {
-    "babel": "^5.8.35",
+    "babel": "^5.8.38",
     "babelify": "^6.4.0",
-    "bannerize": "^1.0.2",
-    "bluebird": "^3.2.2",
+    "bannerize": "^1.1.3",
+    "bluebird": "^3.5.1",
     "browserify": "^12.0.2",
-    "browserify-shim": "^3.8.12",
+    "browserify-shim": "^3.8.14",
     "browserify-versionify": "^1.0.6",
-    "budo": "^8.0.4",
+    "budo": "^8.4.0",
     "deepmerge": "^0.2.10",
-    "eslint": "^2.9.0",
-    "glob": "^6.0.3",
-    "global": "^4.3.0",
+    "eslint": "^2.13.1",
+    "glob": "^6.0.4",
+    "global": "^4.3.2",
     "husky": "^0.14.3",
-    "karma": "^0.13.19",
-    "karma-chrome-launcher": "^0.2.2",
-    "karma-detect-browsers": "^2.0.2",
+    "karma": "^0.13.22",
+    "karma-chrome-launcher": "^0.2.3",
+    "karma-detect-browsers": "^2.2.6",
     "karma-firefox-launcher": "^0.1.7",
     "karma-ie-launcher": "^0.2.0",
     "karma-qunit": "^0.1.9",
     "karma-safari-launcher": "^0.1.1",
     "mkdirp": "^0.5.1",
-    "node-sass": "^3.4.2",
-    "npm-run-all": "^1.5.1",
-    "qunitjs": "^1.21.0",
-    "rimraf": "^2.5.1",
-    "sinon": "~1.14.0",
-    "uglify-js": "^2.6.1",
+    "node-sass": "^3.13.1",
+    "npm-run-all": "^1.8.0",
+    "qunitjs": "^1.23.1",
+    "rimraf": "^2.6.2",
+    "sinon": "^1.14.1",
+    "uglify-js": "^2.8.29",
     "videojs-languages": "^1.0.0",
-    "videojs-standard": "^4.0.0"
+    "videojs-standard": "^4.0.3"
   },
   "scripts": {
     "prebuild": "npm run clean",
